+++ packages/EasyApiPlatform/bundle/config/definition.php
@@ -25,7 +25,7 @@ return static function (DefinitionConfigurator $definition) {
             ->arrayNode('custom_paginator')
                 ->canBeDisabled()
             ->end()
-            ->arrayNode('return_404_on_post')
+            ->arrayNode('return_not_found_on_read_operations')
                 ->canBeDisabled()
             ->end()
         ->end();
