+++ packages/EasyApiPlatform/tests/Fixture/app/config/packages/disable_return_not_found_on_read_operations/easy_api_platform.php
@@ -6,6 +6,6 @@ namespace Symfony\Component\DependencyInjection\Loader\Configurator;
 use Symfony\Config\EasyApiPlatformConfig;
 
 return static function (EasyApiPlatformConfig $easyApiPlatformConfig): void {
-    $easyApiPlatformConfig->return404OnPost()
+    $easyApiPlatformConfig->returnNotFoundOnReadOperations()
         ->enabled(false);
 };
